smt: show original and sliced programs' pre in cex
before, we only showed the sliced program's pre, which was often
confusing because users had to think about the program with the erroring
assertions removed to compute the pre. now the result corresponds to the
original program.

however, we still also print the sliced program's pre value to let the
user focus on the failing pre with just the assertions in the error
message.