proof rules: better error handling
when proof rules were invoked with the wrong number of arguments, Caesar
would just crash because the implementations of `resolve` assumed the
correct number of arguments already, before type checking was done.

we now do resolving independently of the number of arguments and let the
type checking do that checking later.

additionally, we also now properly handle the case where the first
argument to the omega proof rule is not an identifier.