procs: do not add assumes/asserts if there are no pre/posts
we did this before to more closely correspond to some of our old theorems,
but the added statements do not have any semantic effect.

the only (negative) consequence is misleading verification/error
messages for those introduced statements, which had a span of the whole
(co)proc. thus, we had confusing highlights over the whole proc.