slicing: fix unsoundness with exists-forall and nondet choices
previously, the constraints obtained for nondeterministic choices (at
least one branch must be enabled) lead to the solver returning slices
which did not satisfy the contraints. this was because the exists-forall
encoding went as follows:

  exists slice_vars. forall inputs. !(slice_constraints && vc is
  invalid)

instead, it is now the correct one:

  exists slice_vars. slice_constraints && forall inputs. vc is valid

this prevents the solver from returning a slice that does not satisfy
the slice constraints.