Enable datatype specific lit wrapping
The ability to lit wrap is now a trait that a datatype must implement.
This enables different specialisations per datatype.
The main motivation is to avoid boxing the pair EUReal representation
into a single value before passing it into the lit function. The bool
and real are now wrapped separately.