smt: strengthen defs and optimize for macro finder
we now strengthen func declarations in smt by default by not including
constraints for inputs. this over-specifies the definitions, but since
we're guaranteed by the type system that we don't invoke them in a wrong
way, this should not ever be a problem. the advantage is that this
allows Z3's macro finder to work even on funcs with input constraints.

in addition, we now don't translate funcs that *cannot* be recursive via
fuel, but instead always via the simple axiomatic encoding. this also
helps with not confusing Z3.

attached is a simple regression test, which *requires* that we
strengthen its definition in SMT, and also that we don't encode it using
a fuel encoding.