qelim: fix multiplication with infty, allow exprs
Fixes #105.

Before, we did qelim with operands that could be infinite. We now
disallow that because it leads to unsoundness (see linked issue).

Instead of checking that the LHS operand is a constant, we now check
that it is finite only. Thus, we now allow expressions with variables
inside. As a consequence, quantifier elimination after prob. choices
with variables inside now works.

I also added support for traversal of the algorithm through parenthesis
expressions. Those don't come from vcgen, but only from user-written
code.